tests/test_kernel_controller.py
import unittest
from kernel.core.kernel_controller import KernelController

class TestKernelController(unittest.TestCase):
    
    def setUp(self):
        self.controller = KernelController()
    
    def tearDown(self):
        pass

    def test_kernel_functionality(self):
        result = self.controller.get_status()  # Assuming get_status() exists in KernelController
        self.assertIsInstance(result, dict)  # Ensure it returns a dictionary
        self.assertIn('config', result)  # Check if 'config' is in the returned dictionary
        self.assertIn('peer_nodes', result)  # Check if 'peer_nodes' is in the returned dictionary
        self.assertEqual(result['config'], {})  # Assuming the config is empty
        self.assertEqual(result['peer_nodes'], {})  # Assuming peer_nodes is empty

if __name__ == '__main__':
    unittest.main()

